15 passing tests | fixed param overflow in contract

+++ Truffle/contracts/CharacterEveryDay.sol
@@ -21,24 +21,22 @@ contract CharacterEveryDay {
         voteForCharacter(_character);
     }
 
-    //Does NOT prevent overflow ex: 257 on input becomes 1
-    modifier onlyCharacter(uint8 _character) {
+    modifier onlyCharacter(uint256 _character) {
         require(_character < 256);
         _;
     }
     
-    function voteForCharacter(uint8 _character) public onlyCharacter(_character) {
+    function voteForCharacter(uint256 _character) public onlyCharacter(_character) {
         if ((now - timeOfLastVote) > 1 days) {
             selectCharacter(winningCharacter);
         }
         
-        // require(_character <= 256);
         uint256 todaysAddressHash = uint(keccak256(abi.encodePacked(day, msg.sender)));
         require(addressVotedToday[todaysAddressHash] == false);
         
         characterVotes[_character]++;
         if (characterVotes[_character] > winningCharacterVotes) {
-            winningCharacter = _character;
+            winningCharacter = uint8(_character);
             winningCharacterVotes = characterVotes[_character];
         }
         addressVotedToday[todaysAddressHash] = true;
